+++ src/formatter/summary_formatter_spec.ts
@@ -43,7 +43,7 @@ describe('SummaryFormatter', () => {
             '\n' +
             '1) Scenario: b # a.feature:2\n' +
             `   ${figures.cross} Given a failing step # steps.ts:9\n` +
-            '       error\n' +
+            '       Error: error\n' +
             '\n' +
             '1 scenario (1 failed)\n' +
             '1 step (1 failed)\n' +
@@ -76,7 +76,7 @@ describe('SummaryFormatter', () => {
             '\n' +
             '1) Scenario: c # a.feature:3\n' +
             `   ${figures.cross} Given a failing step # steps.ts:9\n` +
-            '       error\n' +
+            '       Error: error\n' +
             '\n' +
             '1 scenario (1 failed)\n' +
             '1 step (1 failed)\n' +
@@ -215,7 +215,7 @@ describe('SummaryFormatter', () => {
             '\n' +
             '1) Scenario: b (attempt 1, retried) # a.feature:2\n' +
             `   ${figures.cross} Given a flaky step # steps.ts:21\n` +
-            '       error\n' +
+            '       Error: error\n' +
             '\n' +
             '1 scenario (1 passed)\n' +
             '1 step (1 passed)\n' +
@@ -249,13 +249,13 @@ describe('SummaryFormatter', () => {
             '\n' +
             '1) Scenario: b (attempt 2) # a.feature:2\n' +
             `   ${figures.cross} Given a failing step # steps.ts:9\n` +
-            '       error\n' +
+            '       Error: error\n' +
             '\n' +
             'Warnings:\n' +
             '\n' +
             '1) Scenario: b (attempt 1, retried) # a.feature:2\n' +
             `   ${figures.cross} Given a failing step # steps.ts:9\n` +
-            '       error\n' +
+            '       Error: error\n' +
             '\n' +
             '1 scenario (1 failed)\n' +
             '1 step (1 failed)\n' +